[Xtext Contentassist] Show rich hover in content assist (patch applied)
This patch adds the eobject hovers to the content assist. Like in JDT a hover
is displayed next to the content assist window. The hover content is derived
from an additional info object. In case it is an eobject the hover is created
by utilizing the DispatchingEObjectTextHover.

The ConfigurableCompletionProposal  is extended to accept any object instead of
just Strings as an additional info object. It also implements now
ICompletionProposalExtension3 and ICompletionProposalExtension5.

Patch by Christoph Kulla
(https://bugs.eclipse.org/bugs/show_bug.cgi?id=331865)