[xtext] Refactoring: Removed ISelector interface
The ISelector marker interface introduced unnecessary complexity and did not provide much benefit. Instead of generic selection, the IScope exposes a fixed set of well defined queries for the existing use cases in the framework. This reduces the migration effort for existing languages and makes it easier to use with less pitfalls with respect to runtime complexity.

Clients are free to introduce own scope implementations with additional querying capabilities such as prefix or fuzzy name lookup.