[xtext.grammar] Support for simple syntactic predicates in Xtext grammar [xbase] Some syntactic changes in Xbase to get rid of backtracking [xtext.ui] Fix in content assist (see xtext.xtext.ui.tests)
Predicates can be defined in the Xtext grammar by means of the '=>' keyword
that prefixes the actual value that should be parsed, e.g.
IfExpression: 
  'if' '(' condition=Expression ')'
    then=Expression
  ( =>'else'
    else=Expression
  )?;

will guide the parser to assign parse the 'else' eager and thereby ensure
that it will be part of the innermost if-expression.

Xbase syntax changes:
- Casted expressions: expression as type instead of (type)expression
- Closures have to be enclosed in brackets as long as they are not the only parameter of a feature call
- Clean-up in switch expression - condition is mandatory
- Type literals use the C# keyword typeof(Name) instead of Name.class