[xtend2.tests] Smoke and robustness tests [xtext] Some fixes to partial parsing algorithm
Important note:
Make sure to use the launch shortcut for fast tests when xtend2 tests 
are executed locally. Smoke tests may take ages ...

Regarding the partial parsing changes:
The node models could be different for a model
that was created by means of XtextResource#update when
compared to an initially parsed resource.