[serializer] bugfix in how NFAs are created
This fix is of major relevance and I'm surprised it hasn't been
discovered earlier. The code that creates state machines (NFAs) for the
new (and the old) serializer didn't handle alternatives with optional
members correctly. If an alternative has at least one optional member,
the alternative itself becomes optional. This wasn't considered so far.