[serializer/tests] disabled and not supported
This test checks if a model can serialized when it has un-scopable cross
referenced. The old serializer used to delegate to the node model, take
the value from there and continue quietly. 

This behavior can lead to 
- overlooked bugs in one's scoping implementation
- mysterious usage of wrong (old) cross reference 
  values during serialization
- Models which have "successfully" been serialized but can not be
  parsed/linked again.
  
Therefore, I think it makes sense to support this mechanism as soem kind
of error recovery. However, it should at least raise a warning.