[serializer] implemented a fail-early strategy for sequencing
The backtracking algorithm of the semantic sequencer now avoids paths
that can not lead to a successful consumption of all EOBject's values.

To do so, every node of the state machine has knowledge about all
EStructuralFeatures which its subsequent nodes refer to. If an EObject
has unconsumed values within EStructuralFeatures that are not part of
this list, the algorithm will not proceed past this node.