[xbase][lib] Added JavaDoc and case for byte[] to ToStringHelper
Please not that we currently do not prevent infinite recursion. 
The algorithm will fail e.g. for EObjects that have a pointer to their container which in turn holds a list of it's contents.

Same would happen for something like

@Data
class Node {
  Node parent;
  List<Node> children();
}

since Node#toString is actually defined, it would not even be possible
to identify that recursion easily. I'm afraid we have to reconsider the approach.