[validation][bug 400989] Get issue location from parent node
DiagnosticConverterImpl#getLocationData() can only retrieve the
location data (line number, offset, and length) if there is any node
in the parse tree directly corresponding to the semantic object. This
commit changes this by retrieving the location data for the container
object in that case. Additionally there is a fallback which
will return location data (0, 0, 0) to make sure that there will
always be an annotation displayed in the editor.

Change-Id: Iddd2e506d815d05f41f261a67e64425dc4e7f2ce