[validation][bug 400989] Get issue location from parent node
DiagnosticConverterImpl#getLocationData() can only retrieve the
location data (line number, offset, and length) if there is any node
in the parse tree directly corresponding to the semantic object. This
commit changes this by using NodeModelUtils.findActualNodeFor() instead
of NodeModelUtils.getNode(). Additionally there is a fallback which
will return location data (0, 0, 0) to make sure that there will
always be an annotation displayed in the editor.