[debug] open the Xbase editor when somebody clicks on a StackFrame
The tricky situation here is that in the OriginalEditorSelector no
IEditInput is provided but just the simple filename of the .class file.

So what we do here is to scan the StackFrames which are active in the
debugger for one that matches the type-name. Luckily, StackFrames have
pointers to source files.

Change-Id: I14ca5e9a6bc5338ed3d96440657c60132a0cad0f