[deps][bug 401285] use package-import to depend on guava
Xtext and Xtend should use Package-Import declarations to depend on
Guava rather than Require-Bundle declarations. The reason for this is
that the user may want to use a different bundle than the standard
com.google.guava bundle; e.g. com.google.guava.jdk5.

Also Xtext and Xtend should when possible omit the version qualifier
from the dependencies on Guava. The reason being that if Xtext depends
upon a specific Guava version, e.g. version="10.0.1", then users would
not really be able to use any other Guava version, since the major
version is incremented with every release. So if a specific minimum
version of Guava is required it is be best to test against all current
Guava versions and then specify a version range as in e.g.
version="[10.0.1,14.0.0)". These dependencies should
then be updated when new versions of Guava become available.

Change-Id: Iae2140ce606f697d9f5300f39b2e161dd5e02047