[xbase][lib] Reviews signatures of generic functions
+ Minor fix in JavaDoc
+ Fixed declaration of CollectionLiterals#new*Map(Pair<>..)
+ Fixed declaration of Map#transformValues (preliminary)

Maps#transform values was declared as in Guava, but it looked
too restrictive to me thus I relaxed it. A question in the Guava
google group was raised to clarify why the signature looked
as it did.

Non of the changes seemingly did have an impact on the generated
code.

see https://bugs.eclipse.org/bugs/show_bug.cgi?id=380139