[performance] Optimize LazyLinker#doLinkModel()
The performance of LazyLinker#doLinkModel() currently suffers
somewhat due to the fact that shouldCheckParentNode() needs to be
called for every EObject and possibly also due to all the created
node model iterators (also at least one per EObject).

This commit implements an alternative approach: Instead of iterating
over the EObjects the LazyLinker now iterates over the node model
(using a single iterator) and then processes all nodes which have a
CrossReference as their grammar element.

The tricky part is figuring out the semantic object to which the
cross reference belongs to. But that is nicely handled by
NodeModelUtils.findActualSemanticObjectFor().