[performance] Optimize LazyLinker#doLinkModel()
Iterating over the node model instead of the resource's EObject
hierarchy does not really improve the performance of the
LazyLinker. Partly this seems to be due to the fact that the
NodeTreeIterator ends up using a lot of resources.

This commit slightly improves the performance of
LazyLinker#installProxies() and NodeTreeIterator#hasNext().

Change-Id: I6e25cc76f830dc29180ee97504e43605c1cbd38a