[memory] Avoid using the SimpleAttributeResolver
Internally it uses a cache with weak references
to EObjects. Additionally it will register a listener
on these objects in order to proactively invalidate
the cached value of the name feature. Reading the
value of #name is much faster than a lookup in a 
WeakHashMap and more memory efficient, too.

Not a big deal in practice, but the memory for 
strongly referenced objects is now at 95MB (!) 
(with 900 (!!) MB of to-be-collected instances which would
still cause the DynamicResourceClusteringPolicy to hit)
for the org.eclipse.xtend.ide.tests.performance.PerformanceTest

see https://bugs.eclipse.org/bugs/show_bug.cgi?id=407451