[common.types] Allow '.' as separator for nested types
A new method #findTypeByName(name, binaryDelimiter) was
added to the IJvmTypeProvider in order to allow the '.'
as a nested type delimiter.

Type- and ConstructorScopes prefer the '.' as the delimiter
for nested types when asked for a description for a given
element.

Organize imports converts the dollar notation to dot separators.

see https://bugs.eclipse.org/bugs/show_bug.cgi?id=406088

Change-Id: I4b5f687ebb219a9ac08ca13717d8be9acc36a4a4