[xbase] Smoother transition for Java devs
This commit makes it easier for Java devs to get used to Xtend.
In fact, it allows more of Java syntax for nested feature calls
and overload identifiers to be type literals.

[syntax] Allow the '.' as type delimiter in static feature calls

Instead of java::lang::String::valueOf(1) it's now also possible
to use java.lang.String::valueOf(1)

[linking] Link feature calls as type literals

This is still work in progress, but the overall feature
works as expected.

The XMemberFeatureCall java.lang.String.valueOf('')
is semantically equivalent to the variants
String::valueOf(''), java::lang::String::valueOf('')
or java.lang.String::valueOf(''). It compiles to the very
same code.

Intermediate expressions, e.g. java.lang in java.lang.String
are flagged as package fragments and linked to the JVMType String.

[navigation] Hyperlinks are produced properly

Nested types produce individual hyperlinks, e.g.
java.util.Map.Entry produces two links. The first one
links on the region 'java.util.Map' to the class Map, and
the second links the nested type Entry on the region 'Entry'

[findReferences] Filter package fragments from result

The package fragments of fully qualified references
refer to the type, too (otherwise they would hold
proxies which may cause a lot of trouble). These references
are not part of the search result.

[organizeImports] Organize imports replaces :: by .

The organize imports infrastructure replaces all occurrences
of $ and :: by a plain dot.

see bugs:
https://bugs.eclipse.org/bugs/show_bug.cgi?id=403838
https://bugs.eclipse.org/bugs/show_bug.cgi?id=406087

Change-Id: I323923d5f1b2fd506d7d2e603003621b9901d3ec