[xbase] Smoother transition for Java devs
This commit makes it easier for Java devs to get used to Xtend.
In fact, it allows more of Java syntax for nested feature calls
and overload identifiers to be type literals.

[syntax] Allow the '.' as type delimiter in static feature calls

Instead of java::lang::String::valueOf(1) it's now also possible
to use java.lang.String::valueOf(1)

[linking] Link feature calls as type literals

This is still work in progress, but the overall feature
works as expected.

The XMemberFeatureCall java.lang.String.valueOf('')
is semantically equivalent to the variants
String::valueOf(''), java::lang::String::valueOf('')
or java.lang.String::valueOf(''). It compiles to the very
same code.

Intermediate expressions, e.g. java.lang in java.lang.String
are flagged as package fragments and linked to the JVMType String.

see bugs:
https://bugs.eclipse.org/bugs/show_bug.cgi?id=403838
https://bugs.eclipse.org/bugs/show_bug.cgi?id=406087

Remaining TODOs:
- Filter find references for package fragments
- Validate String.class and ensure that the user actually meant
  typeof(String).getClass == Class<Class<String>>
- More tests for organize imports and import validation
- Hypelink navigation
  The complete target expression should be the hyperlink region
  rather than the individual segments
- The hover shows the type of the expression. If the target
  is a declaring type of a static feature, the target's type
  is not Class<String> but rather the linked JvmType java.lang.String

Change-Id: I323923d5f1b2fd506d7d2e603003621b9901d3ec