[xtend batch compiler] don't reuse classloading cache from first phase.
The ClassFinder now caches all results (misses and hits). In the first compile phase there are usually not all classes available, so the misses are wrong for the second phase.

Type resolution happens because of annotation processing, which uses the type scope and that one as a new features resolves the whole import section, which includes not-yet-existing types.