[uri hell] In resource descriptions use platform:/resource instead of platform:/plugin
This is a fix for resources which have been loaded using a platform:/plugin URI
This happens when one resource has absolute references using a platform:/plugin uri and the corresponding ResourceDescriptionManager resolves references in the first phase, i.e. during EObjectDecription computation. EMF's GenModelResourceDescriptionStrategy does so as it needs to call GenModel.reconcile() eagerly.