[encoding] Fixed issue in dirty state handling
The string that was encoding as an input stream
by the external content support was not aware
of the expected encoding thus the bytes could not
be decoded to a valid string afterwards.

The LazyStringInputStream was introduced to adress 
the problem. It keeps a reference to the input string
and converts it only on demand to a byte array. The
XtextResource uses the original string of the 
LazyStringInputStream which also saves some negliable
CPU cycles and spares some memory.

see https://bugs.eclipse.org/bugs/show_bug.cgi?id=416872

Change-Id: I46ee6569a5459e2180699849fc75b2d13ae9ee1c