[xbase][linking] Validate suspiciously overloaded methods
Methods that are overloaded by different receivers
may resolve unexpectedly in certain circumstances.

These situations are detected and flagged with a configurable
warning.

In the following example binds to this.foo rather than
it.foo which may not be the intention of all users in
that situation:

class A {
  def foo(CharSequence chars) {}
}
class B {
  def foo(String str) {}
  def bar() {
    new A => [
      foo("baz")
    ]
  }	
}

Navigation allows to go to both candidates: the actually
linked feature and the rejected feature.

see https://bugs.eclipse.org/bugs/show_bug.cgi?id=397062

Change-Id: Iaa83d9aacc3061a9cc09ba3477ecbbd439ea3721