[xtend][parser] Re-enabled partial parsing selectively
Partial parsing is used if the new token sequence
is equal to the old sequence of tokens thus the parser
would use the same lookahead as before. This allows to
use partial parsing whenever the content of a token
(or number of tokens) is edited, e.g. the value of an
ID, a String or a template expression is changed.

see https://bugs.eclipse.org/bugs/show_bug.cgi?id=424804

Change-Id: Ie5a3d6531393047221c8edf76d4bb767376d8c85