[xbase][annotations] Refactored grammar rules
This reduces the required lookahead and improves
the error message. Parsing performance for
annotation lists improves slightly (25% faster
for nested annotation values).

I added AnnotationValueParserBenchmark to show some
numbers on the evolution of the Xtend parser wrt
annotation values.

This change introduces also a more concise notation
for annotation value lists that are assigned
to the default operation 'value', e.g. it is possible
to write @SuppressWarnings('a', 'b') rather than
@SuppressWarnings(#['a', 'b'])

Change-Id: I4497bbb01d744f2b1bd04046701f463a60fbfcf7