[bug 427544][standalone builder]: Fixed issue with onDemand==true.
If onDemand is set to true, it will now first try with onDemand==false, then check for the last segment and only then call super.getResource(uri, true). It will potentially cause trouble, when there are multiple different resources with the same simple name. One could check that by trying to resolve the URI before looking for the last segment.