[xbase][contentAssist] Don't propose #values etc in switch cases
Switch expressions add a local static import if the 
switch value is an enum type. This allows to use the 
enum literals similar to Java. Previously, all static
members of the enum type were added to the static scope
which produced bogus content proposals. This commit
introduces a means to add only members with certain
names. It therefore prepares named static imports, too.
This also paves the way for local imports by annotations
or macros.