[common.types][findReferences] Use monitor less often
It turns out that 'Find references in EMF resources'
during find references is only visible that long because
we use a progress monitor and the number of work items
is equal to the number of resources in the index. With
70 resources, this makes up for more than 2 secs even
if the index query is completely skipped. Firing events
now only after each 10th resource greatly improves the
perceived performance.

Change-Id: I1f8e1dd62ae3e217a4dd38a6e7b2a00f16d85f90