[xtext][junit] Introduced ParallelXtextRunner
A runner that executes the methods of a class
in parallel. There are some restrictions when
the runner can be used, but if these are fulfilled
the test executations speeds up by approx 40%

This is experimental and does not play well with
plugin tests, yet.

Problem is, that plugin tests may cause a lazy bundle
load of UI bundles (via ResourceServiceProviderRegistry)
which has to happen in the UI thread.

Change-Id: I8620cf3474acf50cc7fc61e9db19731ebb2d8689