[xtext][xtext] Fixed broken test on server
Do not unload generated packages since they cannot
be referenced in the same resource set anyway and it
caused trouble: An edit to a GrammarResource triggered
proxification of the generated package and when the
serializer tried to persist the grammar resource, the
proxies would have been resolved again but there is no
physical resource for the generated resource URI
which will put a dummy XMIResource into the resource
set. Afterwards, that would be broken.

Change-Id: Ida5a7b8edbaf3a1d5914a795c0afa90ff5a89fcc