[xtext document] improved cancellation of xtext document clients when new changes have been made.
An adapter on the resource denotes whether the state is outdated. It is set to outdated as soon as document changes come in and set to not outdated when the modify lock has been aquired. As no readers and listeners get called on outdated state the "adapter was removed" exception shouldn't happen anymore. A canceled Typeresolution in general leaves a broken resource behind. Maybe we should clean it up on OperationCanceledExceptions, but it shouldn't be neccessary as it is no longer leaked and cleaned up on reparse.

Change-Id: Icc5d9e52337dd628ad467d15ff924b4ecfcfe602