[xtext][junit4] Attempt to deal with flaky tests
Some of our tests seem to depend on index data to
be present or some other unpredictable things. The
JUnit TestRule Flaky.Rule tries to deal with those
symptomatically. By annotating a test method as
@Flaky, one can indicate that the test should be
executed again if it failed. The number of trials
can be configured and defaults to 3.

see https://bugs.eclipse.org/bugs/show_bug.cgi?id=442349

Change-Id: Ic405b64b3c5c46350fd8a954d6791bf65978a732