[443131] Fixing the name replies by UIStrings class for type references.
The UIStrings class is able to reply the human-readable string for a
JvmTypereference.
When the JvmTypeReference is resolved, the simple name of the referred
type is always replied.
But, when the JvmTypeReference is not resolved, the replied name is
directly extracted from the text of associated code.
Consequently, the replied value may be a simple name or a fully
qualified name of the type.
This patch ensures that UIStrings class always reply the simple name of
the type in all cases.

see https://bugs.eclipse.org/bugs/show_bug.cgi?id=443131

Signed-off-by: sgalland <galland@arakhne.org>