fixed 446319: Avoid NPE in editors with non-matching file extension
The DirtyStateEditorSupport should only announce resources as dirty state,
if the resource's language is the same as the lanugage registered for 
the resource's URI. 

E.g. If you open an "Arithmetics Editor" on "foo.xtext", the edtor sould
not contribute dirty state because its exportet IResourceDescription
can be significalty different (based on the arithmetics language) compared to
what clients expect (IResourceDescription based on the Xtext language).