[xtend] Do go to declaration on F3. Streamlined hyperlinking hovers. (see https://bugs.eclipse.org/bugs/show_bug.cgi?id=443875)
I added 'Open Declaration - ...' to the go to declaration hovers again, as it was quite confusing in some cases to just see the regular label of the referenced element. So now it is always clear that it is the declaration hyper link, but also shows the additional resolved information (i.e. types)

Change-Id: I3aff4867f3e9cee3ac7f1e12360f3592f50d86ad