[xbase][scoping] Make FeatureScopes easier to customize
Clients can now override addToList() on the individual
session based scopes to easily filter the content of each scope.
All feature scopes are created by proper factory methods in the
FeatureScopes class thus the not-so-common-but-existing use case
of specializing the scope content is supported in a more
straight forward way. Since it is a scenario that occurs rarely,
there is no injectable service yet, that customizes the aspects
that are common to all AbstractSessionBasedScopes. This may be
added in the future if we know more about the real world customizing
in that area.

see https://bugs.eclipse.org/bugs/show_bug.cgi?id=454238

Change-Id: I5a9cf7b47874e6a3a65d3bcff2d83141c895fbc3