[serializer] fixes 457579: Method createSequence exceeds bytes limit
This changes the SerializerFragment to generate if-statements
inside AbstractSemanticequencer#createSequence *only* if there 
is more then one scenario to consider.

If there is only one scenario, the only purpose of the if-statements
was to validate internal serializer state. In this regard, the 
serializer implementation is quite stable, so it should be safe to 
omit this validation.

---- old generated code ----
case DummyLangPackage.MODEL:
  if(context == grammarAccess.getModelRule()) {
    sequence_Model(context, (Model) semanticObject); 
    return; 
  }
  else break;
--------
			
---- new generated code ----			
case DummyLangPackage.MODEL:
  sequence_Model(context, (Model) semanticObject); 
  return; 
--------

Change-Id: I2e37e7931343b2492fcb8351b255675dbca34ac6