[batchcompiler] make creating/destroying classloaders configurable
While testing with Gradle, I found out that URLClassLoaders will keep
their Jars open *even after* they were closed. There is a cache in the
JarFileFactory that still references them. There is no way to turn off
the cache. 

Clearing that cache requires lots of ugly reflection code, which I currently do in the Gradle plugin. It felt too risky to include in xtend.core so shortly before the release.

After 2.8, we should investigate how to deal with this properly.