[465080] Unpack classpath containers when scanning projects
During classpath URI resolution, we may need to check project
roots if the file is not in a source folder but will be packaged.
Therefore we need to check the current project and all referenced
projects for a resource in the project root. To determine the
referenced projects, we need to unfold classpath containers like
the plugin classpath container.

see https://bugs.eclipse.org/bugs/show_bug.cgi?id=465080