xtend.web, originally by Huibo Shi, cleaned up by Michael Vorburger
Add auto-edit for the parenthesis marker, bracket marker, accolade
marker, quote market and single quote marker.

Redesign the interface (add the two signal buttons and a combo-box for
selecting example).

Add the xtend examples.

Solve the generation problem when the xtend file contains severals class
by adding a comboBox in right side of the site to save the generate java
file list.

Implement the functionalities when we change the select of the comboBox
and save the changes if users do changes in current session.

Added the second Xtend editor for active annotation. And implement the
action to make the editors collapsible.

improve the collapsible stuff

Separate the java script from html file

xtend-gen on .gitignore && git rm -r xtend-gen/

Signed-off-by: Huibo SHI <shi@itemis.de>
Signed-off-by: Michael Vorburger <mike@vorburger.ch>