Add the xtend examples.
Solve the generation problem when the xtend file contains severals class
by adding a comboBox in right side of the site to save the generate java
file list.

Implement the functionalities when we change the select of the comboBox
and save the changes if users do changes in current session.