[web/formattingtest] fixed test
This fixes a test that was broken by the change linked below. 

The test broken because now the to-be-formatted region also includes the
whitespace that reaches until (but not after) the the end of the
to-be-formatted region. 

This causes this test to also format behind the "end" keyword, as it was
followed by an empty hidden region (zero whitespace characters). This
change adds one whitespace character, causing hidden region to be
outside of the to-be-formatted region.

see https://github.com/eclipse/xtext/pull/819

Signed-off-by: Moritz Eysholdt <moritz.eysholdt@itemis.de>