[427909]: Expose specialized StringConcatenation append methods
Current append(Object) and related methods need perform
instanceof checks to identify the type of object being passed
in -- leading to simple string literals to need two checks
and a call to String.toString().

Provide overloaded versions of append() which take String,
StringConcatenation or StringConcatenationClient, which allows
the compiler to use them directly when the callsite can be statically
determined to pass one of these types.

Signed-off-by: Robert Varga <nite@hq.sk>