[492072] Fixed compiler issue for feature calls
prepareExpression(XExpression,ITreeAppendable):
all feature calls that did not match a condition in the if case should
be handled like the else case. In Xtend there was an issue if the
feature call is for a RichString receiver, which did not require a
variable declaration anymore. In this case the if-block was processed,
while previously it was processed in the else block since variables were
alwas declared before.

Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>