[475431] Provide better error reporting on initialization failure
When initialization fails during creation of the Guice injector, an
analysis of the installed bundles and versions is performed. This is
done to create a detailed error report which should help to identify the
issue.

Changed Access and ExecutableExtensionFactory to be better prepared to
the fact that the bundle was not properly started and injector might be
unavailable.

Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>