[eclipse/xtext-eclipse#299] Added support for Language Server Jar
The IDE project's build can be extended to create an executable Language
Server. The Language Server can be configured to be built either as a
distribution with seperate jars or as a fat jar.
Due to a limitation of the maven-shade-plugin in a Maven/Tycho
environment the build must be tweaked a bit to materialize dependencies
with the maven-dependencies-plugin and attached for inclusion with the
addjars-maven-plugin.

Extended the test suite for the different configurations with language
server support:
- plain Maven
- Maven/Tycho
- Gradle

Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>