[#555] Fix for issue.
Since windows line endings are replaced by unix line endings - i.e. the
"\r" character is removed -, the result string can be shorter than the
input string.
Since the loop iterating over the characters of the input string is
limited to "maxLength + 1" iterations, the check "result.size() >
maxLength" did not give the desired result in the presence of windows
line endings.

Now, the number of iterations is incremented when a character is deleted
from the string.


Signed-off-by: Florian Stolte <fstolte@itemis.de>