[#489] Added choice to select Xtext version
This change offers a drop down box to select "Xtext Version" to use for
p2 install and target platform. The choice will influence the p2 repos
to use for Xtext,MWE,Xpand.
Added Xtext SDK to the target platform for the case that only a subset
of subprojects are installed and required bundles are added from the
target platform then.

Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>