[#462] Remove annotations not based on resource markers.
Xtext adds annotations when validating from the editor. These
annotations are set in addition to the ones based on resource markers.
If an expensive validation is triggered it is necessary to remove these
annotations as well to ensure that validations that are NOT triggered by
expensive validation are not shown (of course they will be re-added in
case the validation is triggered e.g. by an edit again).

Change-Id: I0b2708e8a24a54e382f8a6a809dc8883b99c8637
Signed-off-by: Arne Deutsch <arne@idedeluxe.com>