[#623] Define tycho-surefire-plugin args by profiles
The tycho-surefire-plugin is managed in the parent pom.xml and
configures argLine by a set of properties. The properties are declared
empty and will be redfined by profiles or overriding by system
properties.
Profiles 'macos' and 'jdk9-or-newer' are declared. Profile 'macos' will
add the required '-XstartOnFirstThread' arg and 'jdk9-or-newer' defines
the '--add-modules=ALL-SYSTEM' setting.

Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>