[eclipse/xtext-core#686] Update new project wizard.
Fix case where unsupported JDK is added to the selection combo in case
it is selected as a default. Instead, in the unlikely case that the
default JDK is not supported, the last JDK from the list is selected.

Allow selection of Java 6 and higher in the wizard but warn if it is not
at least java 8.

Remove IDEA from the new project wizard. IDE plugins are not well
maintained any more, no need to bother new users with not working stuff.

Adding "Automatic-Module-Name" to manifest.

Signed-off-by: Arne Deutsch <Arne.Deutsch@itemis.de>