[#681] Fix race condition.
There is no guarantee from the VM that a write to an array from one
thread is visible to any other thread at any time. This will happen
especially when there is a high load with high memory pressure. In this
case I guess the started thread passes the "check.set" but this value
does not be shown to the main thread. That waits sometimes for 40sec in
the while loop until the first threads timed out.

Replaced the code with a CountDownLatch that does not suffer from this
issue. Flakyness should be gone.

Signed-off-by: Arne Deutsch <Arne.Deutsch@itemis.de>