[#681] Clear interrupted state of test tread at start of method.
It looks as if the test runner thread is interrupted. I can't say why.
The test is failing after only 4ms of execution time. Looks like the
test thread is interrupted at the start of the method already. It might
be that we are bitten by a JUnit bug here: junit-team/junit4#1365

Insert a "Thread.currentThread.interrupted" at the start of the test
case to ensure the interrupted flag is cleared.

Signed-off-by: Arne Deutsch <Arne.Deutsch@itemis.de>