[#69] Fix content assist if lexer returns multiple invalid tokens.
If the lexer returns multiple invalid tokens in succession only the last
one is taken into account for content assist. Especially if there are
keywords only (no ID or similar) this is an issue because only the last
character is used as prefix.

The fix assumes that ALL tokens with syntax errors that come in
succession are relevant for content assist. It seems not to be useful to
execute content assist otherwise.

This can never be reproduced with any grammar that includes standart
"Terminals" because of the ID and the ANY_OTHER rule. The ID rule makes
sure that the start is usually read as only ONE token, whereas the
ANY_OTHER rule ensures there is never a lexer error.

Signed-off-by: Arne Deutsch <Arne.Deutsch@itemis.de>