Support JvmCustomAnnotationValue when testing constantExpression of @Inline annotation.
When the @Inline annotation is used into a Xtend or DSL code, and the
field "constantExpression" is specified into this annotation, the Xbase
compiler fails with a class cast exception. This issue is due to the
fact that the value associated "constantExpression" is not of type
JvmBooleanAnnotationValue, but JvmCustomAnnotationValue (with a
XBooleanLiteral inside).

Signed-off-by: Stéphane Galland <galland@arakhne.org>