[#397][#395] Support for ternary expressions and try with resource
* First shot at ternary operator in Xtend
* [#397]  Added validation, qfixes and J2X, if ternary aka conditional expression is used.

If a ternary expression like "a = cond? 1 : 2" is used, an issue is raised. The default setting for this
issue is ignore and can be configured in "Xtend - Errors/Warnings - Code Style".
Also a quickfix is available (if marked as more than ignore), to replace this ternary expression with a
normal if expression: "a = if (cond) 1 else 2". This quickfix is applied recursively to nested usages of
ternary expressions (like "a = cond? !cond? 4 : 5 : 2") . The issue however is raised only once for that line.
Issues and Quickfixes are covered with JUnitTests.

Java2Xtend functionality: When "Paste Java Code" is clicked and if the issue is marked as
ignore (e.g. ternary expressions are allowed), a ternary expression is inserted. If the issue is marked
as info/warning/error, the ternary expression will be replaces with a normal if expression. This, too,
is covered with a small JUnitTest.

Finally also the XtendFormatter supports ternary expressions.

* [#395] Support try with resources

Validate, that a given resources implements the AutoClosable Interface
and that mark as an error otherwise.

Also support content proposals for defined resources.

Finally check the whole functionality with JUnit Tests.
Both, the Error marking, as well as the compilation to correct Java Code is
checked for Java 8 and for Java versions, that do not support try with
resources.

Signed-off-by: Eva Poell <epoell@itemis.com>