[eclipse/xtext#1425] show number of matches in ReferenceSearchViewPage
The Find-References page now shows number of matches as a suffix in the
label that describes the search result.
The label is refreshed when the search finished - not after every single
hit was added to search result to avoid extensive redraws.
The label is also refreshed whenever the user removes a match.

Removed matches are remembered by the SearchResult. When the search is
re-run from the search-history, previously removed matches are not added
again.


Signed-off-by: Mathias Rieder <mathias.rieder@gmail.com>